rework: Organize websocket backend structure
